arquivos/Banco-de-dados/requirements.txt
streamlit
pandas
plotly
altair
psycopg2-binary
python-dotenv
selenium
webdriver-manager
